Move directory ABC to new ims.directory module.
