+++ src/ims/directory/__init__.py
@@ -15,19 +15,15 @@
 ##
 
 """
-Duty Management System integration.
+Incident Management System directory service integration.
 """
 
-from ._directory import DMSDirectory, DMSUser, hashPassword, verifyPassword
-from ._dms import DMSError, DatabaseError, DutyManagementSystem
+from ._directory import IMSDirectory, IMSGroupID, IMSUser, IMSUserID
 
 
 __all__ = (
-    "DatabaseError",
-    "DMSDirectory",
-    "DMSError",
-    "DMSUser",
-    "DutyManagementSystem",
-    "hashPassword",
-    "verifyPassword",
+    "IMSDirectory",
+    "IMSGroupID",
+    "IMSUser",
+    "IMSUserID",
 )
