+++ src/ims/directory/test/test_directory.py
@@ -15,19 +15,7 @@
 ##
 
 """
-Duty Management System integration.
+Tests for L{ims.directory._directory}.
 """
 
-from ._directory import DMSDirectory, DMSUser, hashPassword, verifyPassword
-from ._dms import DMSError, DatabaseError, DutyManagementSystem
-
-
-__all__ = (
-    "DatabaseError",
-    "DMSDirectory",
-    "DMSError",
-    "DMSUser",
-    "DutyManagementSystem",
-    "hashPassword",
-    "verifyPassword",
-)
+__all__ = ()
